requirements.txt
pandas==2.1.4
scikit-learn==1.4.0
joblib==1.3.2
numpy==1.26.2
python-dotenv==1.0.0
flask
mysql-connector-python==8.0.33
pymysql==1.0.3
SQLAlchemy==2.0.23
pulp==2.7.0
shap==0.44.0
